numpy an optional requirement
