+++ .travis.yml
@@ -13,7 +13,7 @@ python:
   - "nightly"
 install:
   - "pip install ."
-  - "pip install -r requirements.txt"
+  - "pip install -r requirements_dev.txt"
 script: py.test --cov haralyzer tests/
 after_success:
   - coveralls
