--- developer_requirements.txt
@@ -1,2 +1,0 @@
-pytest
-bumpversion
