+++ haralyzer/__init__.py
@@ -1,4 +1,7 @@
 """
 Module for analyzing web pages using HAR files
 """
-from .assets import HarParser, MultiHarParser, HarPage
+from .assets import HarParser, HarPage
+
+
+from .multihar import MultiHarParser
